requirements.txt
openrouter
python-telegram-bot
dotenv